Move `app/queries` to `app/model/queries`
They are too specific to live within the `app` folder. It is preferred
to have them within the model folder (close to our domain model)
